Change dependency name
@khink
Could have a look at this? I did not find any reference to where the types import step might be coming from, it is just unknown in a new site. Because the dependencies could not be found, this step is run at the very end, so it would not fail. Did you actually have a types step when you imported this? Maybe we need to add a condition or something. With typeinfo it works in a fresh plone as expected